requirements.txt
torch==2.4.0
diffusers==0.30.3
transformers==4.44.2
accelerate==0.34.2
safetensors
pillow